requirements.txt
asgiref==3.4.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.12
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.8.3
iniconfig==1.1.1
jsonpath==0.82
Markdown==3.3.6
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.8
zipp==3.6.0
